core/requirements.txt
sqlmodel==0.0.8
feedparser==6.0.*
loguru==0.7.*
python-dateutil==2.8.2
requests==2.31.0
alembic==1.13.1
-r local-requirements.txt
